[issue214] implemented comments from code review

+++ src/search/packed_state_entry.h
@@ -1,8 +1,7 @@
 #ifndef PACKED_STATE_ENTRY_H
 #define PACKED_STATE_ENTRY_H
 
-// TODO Make this dependent on the size of a word (32/64 bit).
-typedef unsigned int PackedStateEntry;
+typedef unsigned long PackedStateEntry;
 
 
 #endif
